refactor(sim): extract shared test fixture factories into test-helpers.ts
Hoists duplicated factory functions (makeAttributes, makeFingerprint,
makePlayer, makeCoachingMods, makeRng, makeSituation, makeStarters) from
simulate-game.test.ts, resolve-play.test.ts, and resolve-matchups.test.ts
into a shared test-helpers.ts module. This eliminates ~80 lines of
duplicated test setup code and ensures a single source of truth for test
fixture defaults, preventing silent drift between test suites.

Closes #377
